LayoutTests/platform/blackberry/fast/encoding/idn-security-expected.txt

FAIL testIDNRoundTrip(0x2e) should be .. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2e) should be .. Was .
FAIL testIDNRoundTrip(0x2f) should be /. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2f) should be /. Was .
FAIL testIDNRoundTrip(0x61) should be a. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x61) should be a. Was .
FAIL testIDNRoundTrip(0x633) should be %u0633. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x633) should be %u0633. Was .
FAIL testIDNRoundTrip(0x561) should be %u0561. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x561) should be %u0561. Was .
FAIL testIDNRoundTrip(0x3105) should be %u3105. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3105) should be %u3105. Was .
FAIL testIDNRoundTrip(0x1613) should be %u1613. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x1613) should be %u1613. Was .
FAIL testIDNRoundTrip(0x905) should be %u0905. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x905) should be %u0905. Was .
FAIL testIDNRoundTrip(0xa85) should be %u0A85. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xa85) should be %u0A85. Was .
FAIL testIDNRoundTrip(0xa05) should be %u0A05. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xa05) should be %u0A05. Was .
FAIL testIDNRoundTrip(0x1115) should be %u1115. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x1115) should be %u1115. Was .
FAIL testIDNRoundTrip(0x4e2d) should be %u4E2D. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x4e2d) should be %u4E2D. Was .
FAIL testIDNRoundTrip(0x5d0) should be %u05D0. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x5d0) should be %u05D0. Was .
FAIL testIDNRoundTrip(0x3041) should be %u3041. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3041) should be %u3041. Was .
FAIL testIDNRoundTrip(0x30a1) should be %u30A1. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x30a1) should be %u30A1. Was .
FAIL testIDNRoundTrip(0xb94) should be %u0B94. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xb94) should be %u0B94. Was .
FAIL testIDNRoundTrip(0xe01) should be %u0E01. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xe01) should be %u0E01. Was .
FAIL testIDNRoundTrip(0xa000) should be %uA000. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xa000) should be %uA000. Was .
FAIL testIDNRoundTrip(0x2024) should be .. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2024) should be .. Was .
FAIL testIDNRoundTrip(0xfe52) should be .. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe52) should be .. Was .
FAIL testIDNRoundTrip(0xff0f) should be /. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xff0f) should be /. Was .
FAIL testIDNRoundTrip(0xfe68) should be %5C. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe68) should be %5C. Was .
FAIL testIDNRoundTrip(0xff3c) should be %5C. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xff3c) should be %5C. Was .
FAIL testIDNRoundTrip(0xa0) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xa0) should be %20. Was .
FAIL testIDNRoundTrip(0x2000) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2000) should be %20. Was .
FAIL testIDNRoundTrip(0x2001) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2001) should be %20. Was .
FAIL testIDNRoundTrip(0x2002) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2002) should be %20. Was .
FAIL testIDNRoundTrip(0x2003) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2003) should be %20. Was .
FAIL testIDNRoundTrip(0x2004) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2004) should be %20. Was .
FAIL testIDNRoundTrip(0x2005) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2005) should be %20. Was .
FAIL testIDNRoundTrip(0x2006) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2006) should be %20. Was .
FAIL testIDNRoundTrip(0x2007) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2007) should be %20. Was .
FAIL testIDNRoundTrip(0x2008) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2008) should be %20. Was .
FAIL testIDNRoundTrip(0x2009) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2009) should be %20. Was .
FAIL testIDNRoundTrip(0x200a) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x200a) should be %20. Was .
FAIL testIDNRoundTrip(0x202f) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x202f) should be %20. Was .
FAIL testIDNRoundTrip(0x205f) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x205f) should be %20. Was .
FAIL testIDNRoundTrip(0x3000) should be %20. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3000) should be %20. Was .
FAIL testIDNRoundTrip(0xbc) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xbc) should be punycode. Was .
FAIL testIDNRoundTrip(0xbd) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xbd) should be punycode. Was .
FAIL testIDNRoundTrip(0xed) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xed) should be punycode. Was .
FAIL testIDNRoundTrip(0x1c3) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x1c3) should be punycode. Was .
FAIL testIDNRoundTrip(0x251) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x251) should be punycode. Was .
FAIL testIDNRoundTrip(0x261) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x261) should be punycode. Was .
FAIL testIDNRoundTrip(0x337) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x337) should be punycode. Was .
FAIL testIDNRoundTrip(0x337) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x337) should be punycode. Was .
FAIL testIDNRoundTrip(0x338) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x338) should be punycode. Was .
FAIL testIDNRoundTrip(0x338) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x338) should be punycode. Was .
FAIL testIDNRoundTrip(0x5b4) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x5b4) should be punycode. Was .
FAIL testIDNRoundTrip(0x5bc) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x5bc) should be punycode. Was .
FAIL testIDNRoundTrip(0x660) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x660) should be punycode. Was .
FAIL testIDNRoundTrip(0x6f0) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x6f0) should be punycode. Was .
FAIL testIDNRoundTrip(0x115f) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x115f) should be punycode. Was .
FAIL testIDNRoundTrip(0x1160) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x1160) should be punycode. Was .
FAIL testIDNRoundTrip(0x2027) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2027) should be punycode. Was .
FAIL testIDNRoundTrip(0x2039) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2039) should be punycode. Was .
FAIL testIDNRoundTrip(0x203a) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x203a) should be punycode. Was .
FAIL testIDNRoundTrip(0x2044) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2044) should be punycode. Was .
FAIL testIDNRoundTrip(0x2044) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2044) should be punycode. Was .
FAIL testIDNRoundTrip(0x2154) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2154) should be punycode. Was .
FAIL testIDNRoundTrip(0x2155) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2155) should be punycode. Was .
FAIL testIDNRoundTrip(0x2156) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2156) should be punycode. Was .
FAIL testIDNRoundTrip(0x2159) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2159) should be punycode. Was .
FAIL testIDNRoundTrip(0x215a) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x215a) should be punycode. Was .
FAIL testIDNRoundTrip(0x215b) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x215b) should be punycode. Was .
FAIL testIDNRoundTrip(0x215f) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x215f) should be punycode. Was .
FAIL testIDNRoundTrip(0x2215) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2215) should be punycode. Was .
FAIL testIDNRoundTrip(0x2216) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2216) should be punycode. Was .
FAIL testIDNRoundTrip(0x233f) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x233f) should be punycode. Was .
FAIL testIDNRoundTrip(0x23ae) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x23ae) should be punycode. Was .
FAIL testIDNRoundTrip(0x244a) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x244a) should be punycode. Was .
FAIL testIDNRoundTrip(0x2571) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2571) should be punycode. Was .
FAIL testIDNRoundTrip(0x2572) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2572) should be punycode. Was .
FAIL testIDNRoundTrip(0x29f6) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x29f6) should be punycode. Was .
FAIL testIDNRoundTrip(0x29f8) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x29f8) should be punycode. Was .
FAIL testIDNRoundTrip(0x29f8) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x29f8) should be punycode. Was .
FAIL testIDNRoundTrip(0x2afb) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2afb) should be punycode. Was .
FAIL testIDNRoundTrip(0x2afd) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x2afd) should be punycode. Was .
FAIL testIDNRoundTrip(0x3014) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3014) should be punycode. Was .
FAIL testIDNRoundTrip(0x3015) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3015) should be punycode. Was .
FAIL testIDNRoundTrip(0x3033) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3033) should be punycode. Was .
FAIL testIDNRoundTrip(0x3035) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3035) should be punycode. Was .
FAIL testIDNRoundTrip(0x3164) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3164) should be punycode. Was .
FAIL testIDNRoundTrip(0x321d) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x321d) should be punycode. Was .
FAIL testIDNRoundTrip(0x321e) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x321e) should be punycode. Was .
FAIL testIDNRoundTrip(0x33ae) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x33ae) should be punycode. Was .
FAIL testIDNRoundTrip(0x33af) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x33af) should be punycode. Was .
FAIL testIDNRoundTrip(0x33c6) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x33c6) should be punycode. Was .
FAIL testIDNRoundTrip(0x33df) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x33df) should be punycode. Was .
FAIL testIDNRoundTrip(0xfe14) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe14) should be punycode. Was .
FAIL testIDNRoundTrip(0xfe15) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe15) should be punycode. Was .
FAIL testIDNRoundTrip(0xfe3f) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe3f) should be punycode. Was .
FAIL testIDNRoundTrip(0xfe5d) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe5d) should be punycode. Was .
FAIL testIDNRoundTrip(0xfe5e) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xfe5e) should be punycode. Was .
FAIL testIDNRoundTrip(0xffa0) should be punycode. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xffa0) should be punycode. Was .
FAIL testIDNEncode(0x2028) should be %u2028. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2028) should be %u2028. Was .
FAIL testIDNEncode(0x2029) should be %u2029. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2029) should be %u2029. Was .
FAIL testIDNEncode(0x2ff0) should be %u2FF0. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff0) should be %u2FF0. Was .
FAIL testIDNEncode(0x2ff1) should be %u2FF1. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff1) should be %u2FF1. Was .
FAIL testIDNEncode(0x2ff2) should be %u2FF2. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff2) should be %u2FF2. Was .
FAIL testIDNEncode(0x2ff3) should be %u2FF3. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff3) should be %u2FF3. Was .
FAIL testIDNEncode(0x2ff4) should be %u2FF4. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff4) should be %u2FF4. Was .
FAIL testIDNEncode(0x2ff5) should be %u2FF5. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff5) should be %u2FF5. Was .
FAIL testIDNEncode(0x2ff6) should be %u2FF6. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff6) should be %u2FF6. Was .
FAIL testIDNEncode(0x2ff7) should be %u2FF7. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff7) should be %u2FF7. Was .
FAIL testIDNEncode(0x2ff8) should be %u2FF8. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff8) should be %u2FF8. Was .
FAIL testIDNEncode(0x2ff9) should be %u2FF9. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ff9) should be %u2FF9. Was .
FAIL testIDNEncode(0x2ffa) should be %u2FFA. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ffa) should be %u2FFA. Was .
FAIL testIDNEncode(0x2ffb) should be %u2FFB. Was .
FAIL testIDNEncodeNotFirstCharacter(0x2ffb) should be %u2FFB. Was .
FAIL testIDNEncode(0xfff9) should be %uFFF9. Was .
FAIL testIDNEncodeNotFirstCharacter(0xfff9) should be %uFFF9. Was .
FAIL testIDNEncode(0xfffa) should be %uFFFA. Was .
FAIL testIDNEncodeNotFirstCharacter(0xfffa) should be %uFFFA. Was .
FAIL testIDNEncode(0xfffb) should be %uFFFB. Was .
FAIL testIDNEncodeNotFirstCharacter(0xfffb) should be %uFFFB. Was .
FAIL testIDNEncode(0xfffc) should be %uFFFC. Was .
FAIL testIDNEncodeNotFirstCharacter(0xfffc) should be %uFFFC. Was .
FAIL testIDNEncode(0xfffd) should be %uFFFD. Was .
FAIL testIDNEncodeNotFirstCharacter(0xfffd) should be %uFFFD. Was .
FAIL testIDNRoundTrip(0x5c3) should be punycode. Was .
FAIL testIDNEncodeNotFirstCharacter(0x5c3) should be %u05C3. Was .
FAIL testIDNRoundTrip(0x5f4) should be punycode. Was .
FAIL testIDNEncodeNotFirstCharacter(0x5f4) should be %u05F4. Was .
FAIL testIDNRoundTrip(0x6d4) should be punycode. Was .
FAIL testIDNEncodeNotFirstCharacter(0x6d4) should be %u06D4. Was .
FAIL testIDNRoundTrip(0x702) should be punycode. Was .
FAIL testIDNEncodeNotFirstCharacter(0x702) should be %u0702. Was .
FAIL testIDNEncode(0x200b) should be %u200B. Was .
PASS testIDNRoundTripNotFirstCharacter(0x200b) is ''
FAIL testIDNEncode(0x3002) should be %u3002. Was .
FAIL testIDNRoundTripNotFirstCharacter(0x3002) should be .. Was .
FAIL testIDNEncode(0xff0e) should be %uFF0E. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xff0e) should be .. Was .
FAIL testIDNEncode(0xff61) should be %uFF61. Was .
FAIL testIDNRoundTripNotFirstCharacter(0xff61) should be .. Was .
FAIL testIDNEncode(0xfeff) should be %uFEFF. Was .
PASS testIDNRoundTripNotFirstCharacter(0xfeff) is ''
